Test Suite AllTests created

+++ src/AdminFunctionalities/AdminFunctionalityController.java
@@ -2,12 +2,10 @@ package AdminFunctionalities;
 
 //imports
 import java.util.*;
-
 import AccountFunctionalities.Account;
 import CMCDatabase.*;
 import UniversityFunctionalities.*;
 import UserFunctionalities.*;
-//import other.*;
 
 /**
  * @author CtrlAltElite
@@ -18,19 +16,20 @@ public class AdminFunctionalityController {
 	private DBController database;
 	private UniversityController uniController;
 	private UserFunctionalityController ufc;
-//	private Account acc;
 	
 	//constructor
 	public AdminFunctionalityController() {
+		
 	database = new DBController();
 	uniController = new UniversityController();
 	ufc = new UserFunctionalityController();
 	}
 	
+	
+	
 	/**
 	 * adds a university to a User's saves schools list
 	 */
-
 	public boolean addToSavedSchoolsList3(String username, String schoolName){
 		
 		DBController dbc = new DBController();
@@ -39,6 +38,7 @@ public class AdminFunctionalityController {
     	int i = 0;  		
     	boolean e = false;
     	outerloop:
+    		
     	while(i<dbc.loadUsers().size()) {
     		String name = dbc.loadUsers().get(i).getEmail().toUpperCase();
    			
@@ -51,31 +51,18 @@ public class AdminFunctionalityController {
     		if(i == (dbc.loadUsers().size())-1) {
 		   		System.out.print(userToFind + " does NOT exist.");
     		}
+    		
    			i++;    			
    		}
-    	
+	
    		if(e) {
    			UniversityController unc = new UniversityController();
    			unc.addToSavedSchoolsList2(userToFind, schoolName);
     	}	
    		return e;
 	}
+
 	
-	
-	
-	
-	
-	/**
-	 * Retrieves universities from database
-	 * by calling loadUniversities() in DBController
-	 */
-	public List<University> viewUniversities() {
-		List<University> listUnis = database.loadUniversities();
-		for(int i = 0; i < listUnis.size(); i++) {
-			System.out.println(listUnis.get(i).getSchoolName());
-		}		
-		return listUnis;
-	}
 	
 	/**
 	 * Adds a new university to the database.
@@ -89,33 +76,15 @@ public class AdminFunctionalityController {
 				percentFemale, verbalSAT, mathSAT, schoolExpenses, percentFinancialAid, numApplicants,
 				percentAdmitted, percentEnrolled, academicScale, socialScale, qualityScale);
 	}
-	
-	/**
-	 * Edits the info stored for a university
-	 * by calling editUniversity() in UniversityController
-	 */
-	public void editUniversity() {
-		uniController.editUniversity();
-	}
-	
-	/**
-	 * Retrieves user info from database 
-	 * by calling loadUsers() on a DBController object and returning it
-	 * @return List of Users
-	 */
-	public List<Account> viewUsers() {
-		List<Account> listUsers = database.loadUsers();
-		for(int i = 0; i < listUsers.size(); i++) {
-			System.out.println(listUsers.get(i).getFirstName() + " " + listUsers.get(i).getLastName() + " " + listUsers.get(i).getEmail() + " " + listUsers.get(i).getStatus() + " " + listUsers.get(i).getType());
-		}	
-		return listUsers;
-	}
+
+
 	
 	/**
 	 * Adds a user to the database
 	 * by getting the information from the user
 	 */
 	public void addUser() {
+		
 		DBController dbc = new DBController();
 		List<Account> users = dbc.loadUsers();
 		
@@ -133,11 +102,14 @@ public class AdminFunctionalityController {
 		char type = sc.next().charAt(0); 
 		boolean o = true;
 		//this.database.checkPasswordRequirements(password);
+		
 		for (int i = 0; i < users.size(); i ++) {
+			
 			if(username.equals(users.get(i).getEmail()) || (this.database.checkPasswordRequirements(password)== false)) {//FINAL PRODUCT WILL NOT HAVE PRINT STATEMENT
 				System.out.println("This username is taken or password does not meet requirements.");	
 				break;
 			}
+			
 			else {
 				dbc.addUser(firstName, lastName, username, password, type);
 			}
@@ -146,30 +118,94 @@ public class AdminFunctionalityController {
 		sc.close();
 	}
 	
+	
+	
 	/**
 	 * Removes university from the database
 	 * by making sure the University is not in the 
 	 * saved schools list of any users
 	 * then calls removeUniversity() in DBController
 	 */
-	public boolean removeUniversityDB(String schoolName) {
-		List<Account> users = database.loadUsers();	
-		boolean success = true;
-		for (int i = 0; i < users.size(); i++) {
-			 Map<String, String> savedList = ufc.getSavedSchoolsList(users.get(i).getEmail());
-			 for (Map.Entry entry : savedList.entrySet()){
-				 if(schoolName.toUpperCase().equals(entry.getKey().toString().toUpperCase())) {
-					 success = false;
-					 System.out.println("School was removed: " + success);
-					 break;
-				 } 
-			 }
+	public boolean removeUniversityDB(String schoolName) {		
+		int j = 0;  		
+		boolean e = false;
+		
+		//searches through list of universities retrieved through loadUniversities() method
+		while(j<database.loadUniversities().size()) {
+			
+			String name = database.loadUniversities().get(j).getSchoolName();
+			
+			//confirms that the uniToFind exists
+			if (name.toUpperCase().equals(schoolName)){
+					e= true;
+					//System.out.println("YES, " + schoolName + " exists.");   			
+					break;
+			}
+			
+			//if the while loop reaches the end of the list, uniToFind does not exist
+			if(j == (database.loadUniversities().size())-1) {
+		   		//System.out.println("School does not exist, did not remove school.");
+				return e;
+			}
+				j++;    			
+			}
+		
+		//if uniToFind exists, 
+			if(e) {
+				
+				List<Account> users = database.loadUsers();	
+				boolean success = true;
+				
+				for (int i = 0; i < users.size(); i++) {
+					 Map<String, String> savedList = ufc.getSavedSchoolsList(users.get(i).getEmail());
+					 
+					 for (Map.Entry entry : savedList.entrySet()){
+						 
+						 if(schoolName.toUpperCase().equals(entry.getKey().toString().toUpperCase())) {
+							 success = false;
+							 System.out.println("School was removed: " + success);
+							 break;
+						 } 
+					 }
+				}
+				
+				if(success) {
+					database.removeUniversityDB(schoolName);
+					System.out.println("School was removed: " + success );
+				}
+				return success;
 		}
-		if(success) {
-			database.removeUniversityDB(schoolName);
-			System.out.println("School was removed: " + success );
-		}
-		return success;
+			return e;
+		
+	}
+	
+	
+	
+	/**
+	 * Retrieves universities from database
+	 * by calling loadUniversities() in DBController
+	 */
+	public List<University> viewUniversities() {
+		List<University> listUnis = database.loadUniversities();
+		for(int i = 0; i < listUnis.size(); i++) {
+			System.out.println(listUnis.get(i).getSchoolName());
+		}		
+		return listUnis;
+	}
+	
+	
+	
+	/**
+	 * Retrieves user info from database 
+	 * by calling loadUsers() on a DBController object and returning it
+	 * @return List of Users
+	 */
+	public List<Account> viewUsers() {
+		List<Account> listUsers = database.loadUsers();
+		for(int i = 0; i < listUsers.size(); i++) {
+			System.out.println(listUsers.get(i).getFirstName() + " " + listUsers.get(i).getLastName() + " " + listUsers.get(i).getEmail() + " " + listUsers.get(i).getStatus() + " " + listUsers.get(i).getType());
+		}	
+		return listUsers;
 	}
 	
 }
